Refactor 'destinations' file structure. Fix broken 'NewDestinationForm'
